requirements.txt
uvicorn
fastapi
streamlit
pandas
openai
pyyaml
pydub
moviepy
botocore
boto3
python-multipart
-e .